wip: refactor to move code under operator
